refactor: Used linq in custom query builder
